User gets dialog on "delete"

+++ src/flb/category/application/listeners/UserDeletesCategoryListener.java
@@ -10,17 +10,24 @@ public class UserDeletesCategoryListener implements ActionListener {
 
     private final CategoryStorage categoryStorage;
     private final CategoryTable categoryTable;
+    private final JFrame frame;
 
-    public UserDeletesCategoryListener(CategoryStorage categoryStorage, CategoryTable categoryTable){
+    public UserDeletesCategoryListener(CategoryStorage categoryStorage, CategoryTable categoryTable, JFrame frame){
         this.categoryStorage = categoryStorage;
         this.categoryTable = categoryTable;
+        this.frame = frame;
     }
 
     @Override
     public void actionPerformed(ActionEvent e) {
         String categoryNameToDelete = categoryTable.getSelectedRowName();
-        categoryStorage.deleteCategory(categoryNameToDelete);
-        ArrayList<Category> categories = categoryStorage.getCategories(categoryTable.getFilterText());
-        categoryTable.rowChangeRefresh(categories);
+        int selection = JOptionPane.showConfirmDialog(
+                frame, "Are you sure you want to delete " + categoryNameToDelete + "?",
+                "Confirm Delete", JOptionPane.YES_NO_OPTION);
+        if (selection == JOptionPane.YES_OPTION) {
+            categoryStorage.deleteCategory(categoryNameToDelete);
+            ArrayList<Category> categories = categoryStorage.getCategories(categoryTable.getFilterText());
+            categoryTable.rowChangeRefresh(categories);
+        }
     }
 }
